+++ expression.hpp
@@ -106,7 +106,7 @@ public:
   bool isText() const noexcept;
   std::tuple<double, double, bool> getPosition() const noexcept;
   std::vector<Expression> asVector() const noexcept;
-  size_t getNumericalProperty(std::string) const noexcept;
+  double getNumericalProperty(std::string) const noexcept;
   std::pair<double, double> getPointCoordinates() const noexcept;
   std::pair<double, double> getTextCoordinates() const noexcept;
